+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Increase modularity in file structure and CSS class definition to improve readability, tracing, etc.
 - Remove all nasty CSS hacks (no more !important here)
+- Changed logging setup for flexibility of info, warn, error, verbosity. Default: Warnings & Errors.